backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
numpy
iqoption==6.8.9.1
websocket-client==0.56.0
requests
webdriver-manager
selenium
MetaTrader5
pydantic
email-validator
scikit-learn
xgboost
SQLAlchemy>=2.0.0
pg8000
cryptography
